+++ CHANGELOG.md
@@ -79,6 +79,13 @@ const updatedSettings = await liveblocks.updateNotificationSettings({
 await liveblocks.deleteNotificationSettings({ userId });
 ```
 
+### `@liveblocks/emails`
+
+- Update the behavior of `prepareThreadNotificationEmailAsHtml` and
+  `prepareThreadNotificationEmailAsReact`: the contents of previous emails data
+  are now taken into account to avoid repeating mentions and replies that are
+  still unread but have already been extracted in another email data.
+
 ## v2.17.0
 
 ### `@liveblocks/client`